refactor(galgame): split plugin_util_helpers.py from __init__.py
Move 6 utility helpers verbatim:
- _log_plugin_noncritical
- _public_context_snapshot
- _migrate_legacy_capture_backend
- _duration_percentile
- _duration_summary
- _open_url_in_browser

__init__.py re-exports them so external import paths stay compatible.
